requirements.txt
flask
ultralytics
opencv-python-headless==4.9.0.80
faiss-cpu
face_recognition